backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0
python-dotenv==1.2.2
boto3==1.42.78
django-storages==1.14.6
Pillow==12.1.1
